split out assets
